Added confirm dialog for deleting airplane records

+++ src/app/pages/airplane-overview/airplane-overview.component.ts
@@ -5,19 +5,41 @@ import { CardModule } from 'primeng/card';
 import { TableModule } from 'primeng/table';
 import { Airplane } from '../../data/airplane';
 import { AirplaneService } from '../../service/airplane.service';
+import { ConfirmPopupModule } from 'primeng/confirmpopup';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { ToastModule } from 'primeng/toast';
 
 @Component({
   selector: 'app-airplane-overview',
-  imports: [TableModule, CardModule, ButtonModule],
+  imports: [
+    TableModule,
+    CardModule,
+    ButtonModule,
+    ConfirmPopupModule,
+    ToastModule,
+  ],
+  providers: [ConfirmationService, MessageService],
   templateUrl: './airplane-overview.component.html',
   styleUrl: './airplane-overview.component.css',
 })
 export class AirplaneOverviewComponent {
   airplaneData: Airplane[] = [];
 
-  constructor(private service: AirplaneService, private router: Router) {}
+  constructor(
+    private service: AirplaneService,
+    private router: Router,
+    private confirmationService: ConfirmationService,
+    private messageService: MessageService
+  ) {}
 
   ngOnInit() {
+    const toast = history.state.toast;
+    console.log(toast);
+
+    if (toast) {
+      this.messageService.add(toast);
+    }
+
     this.loadAirplanes();
   }
 
@@ -37,10 +59,32 @@ export class AirplaneOverviewComponent {
     await this.router.navigate(['airplane']);
   }
 
-  deleteAirplane(airplane: Airplane) {
-    this.service.deleteAirplane(airplane.id).subscribe((obj) => {
-      console.log(obj);
-      this.loadAirplanes();
+  deleteAirplane(airplane: Airplane, event: Event) {
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'Do you want to delete this record?',
+      icon: 'pi pi-info-circle',
+      rejectButtonProps: {
+        label: 'Cancel',
+        severity: 'secondary',
+        outlined: true,
+      },
+      acceptButtonProps: {
+        label: 'Delete',
+        severity: 'danger',
+      },
+      accept: () => {
+        this.service.deleteAirplane(airplane.id).subscribe((obj) => {
+          console.log(obj);
+          this.loadAirplanes();
+          this.messageService.add({
+            severity: 'success',
+            summary: 'Confirmed',
+            detail: 'Record deleted',
+            life: 3000,
+          });
+        });
+      },
     });
   }
 }
